Set max and min clip values dtypes to float32

+++ skrl/models/torch/deterministic.py
@@ -55,8 +55,8 @@ class DeterministicMixin:
         self._d_clip_actions[role] = clip_actions and issubclass(type(self.action_space), gym.Space)
 
         if self._d_clip_actions[role]:
-            self.clip_actions_min = torch.tensor(self.action_space.low, device=self.device)
-            self.clip_actions_max = torch.tensor(self.action_space.high, device=self.device)
+            self.clip_actions_min = torch.tensor(self.action_space.low, device=self.device, dtype=torch.float32)
+            self.clip_actions_max = torch.tensor(self.action_space.high, device=self.device, dtype=torch.float32)
 
             # backward compatibility: torch < 1.9 clamp method does not support tensors
             self._backward_compatibility = tuple(map(int, (torch.__version__.split(".")[:2]))) < (1, 9)
